Fix issue with modularizing samplers
